Extract LigandModal for molecule details
